Add files via upload
Project 1 - We all have played rock, paper scissors game in our childhood. Write a python program capable of playing this game with the 
user.

Project 2 - We are going to write a program that generates a random number and asks the user to guess it. 
If the player’s guess is higher than the actual number, the program displays “Lower number please”. Similarly, if the user’s guess is too low, the program prints “higher 
number please” When the user guesses the correct number, the program displays the number of guesses the player used to arrive at the number.